refactor(views): 抽数据 helpers 与「其他平台」settings tab 到 services/views
Phase 3 第二步。两块抽出：

1. services/wechatsync-settings.js
   把 input.js 顶部 ~165 行 multi-platform sync 数据 helpers 抽成模块：
   - createDefaultMultiPlatformSyncSettings
   - normalizeWechatsyncPlatformId / parseWechatsyncPlatformIds
   - mergeWechatsyncPlatformLists
   - normalizeWechatSyncCapabilities / hasWechatSyncCapability
   - normalizeWechatSyncRecentTasks
   - normalizeMultiPlatformConnection / normalizeMultiPlatformSyncSettings
   - getConfiguredWechatsyncPlatforms / getAvailableWechatsyncPlatforms
   全部纯数据，无 DOM / Obsidian API。views 层不再依赖 input.js（避免循环）。

2. views/settings/multi-platform-tab.js
   把 display() 中 ~430 行的「其他平台」Tab 逻辑全部抽出：
   - 启用 toggle、port / token 输入
   - 持久连接状态条（Phase 2 helper）
   - 平台 picker grid（chip 含 auth 状态）
   - 测试连接按钮（bridge.start + health + listSupportedPlatforms +
     getAuthSnapshot）
   - 诊断已选平台（批量 checkAuth + 单个回退）
   导出 renderMultiPlatformSettingsTab(tab, containerEl)。

input.js 从 7498 → 7068 行 (-430)。51 测试文件 / 594 用例全过。

下一步抽 publish modal showMultiPlatformSyncModal。
